+++ doc/source/releasenotes.rst
@@ -9,7 +9,8 @@ Release notes
 * Introduces a new :class:`~clease.datastructures.TransMatrix` dataclass
   for the translation matrix.
 * Temporarily restricts ASE to ``<3.23``, until we resolve issues with current ASE master.
-
+* Montecarlo will no longer consider background indices in the default swap move generator,
+  if background is ignored.
 
 0.11.1
 =======